fixes in visiond script

+++ selfdrive/car/modules/ch_visiond.sh
@@ -1,14 +1,20 @@
 #!/usr/bin/bash
 cd /data/openpilot/selfdrive/visiond
 if [ "$1" = "wiggly" ]; then
-  ln -sf ./visiond-wiggly ./visiond
-  sleep 2
-  reboot
+  if [ "$(readlink ./visiond)" != "./visiond-wiggly" ]; then
+    ln -sf ./visiond-wiggly ./visiond
+    sleep 2
+    echo "Rebooting"
+    reboot
+  fi
 else
   if [ "$1" = "normal" ]; then
-    ln -sf ./visiond-normal ./visiond
-    sleep 2
-    reboot
+    if [ "$(readlink ./visiond)" != "./visiond-normal" ]; then
+      ln -sf ./visiond-normal ./visiond
+      sleep 2
+      echo "Rebooting"
+      reboot
+    fi
   fi
 fi
